requirements.txt
Django==1.8.1
argparse==1.2.1
fapws3==0.11.dev
python-memcached==1.54
six==1.9.0
wsgiref==0.1.2
